Run again integrationTests on linux
