+++ buildLogic/toolchain-convention/build.gradle
@@ -18,5 +18,4 @@ gradlePlugin {
 
 dependencies {
   compileOnly libs.kotlin.plugin
-  compileOnly libs.android.plugin
 }
