+++ sqldelight-gradle-plugin/build.gradle
@@ -17,11 +17,15 @@ testing {
     grammarkitTest(JvmTestSuite) {
       useJUnit()
     }
+    instrumentationTest(JvmTestSuite) {
+      useJUnit()
+    }
   }
 }
 
 configurations.dockerTestImplementation.extendsFrom(configurations.testImplementation)
 configurations.grammarkitTestImplementation.extendsFrom(configurations.testImplementation)
+configurations.instrumentationTestImplementation.extendsFrom(configurations.testImplementation)
 
 gradlePlugin {
   plugins {
@@ -31,7 +35,7 @@ gradlePlugin {
     }
   }
 
-  testSourceSets(sourceSets.test, sourceSets.dockerTest, sourceSets.grammarkitTest)
+  testSourceSets(sourceSets.test, sourceSets.dockerTest, sourceSets.grammarkitTest, sourceSets.instrumentationTest)
 }
 
 
@@ -72,7 +76,6 @@ tasks.named('test') {
       ":runtime:publishKotlinMultiplatformPublicationToInstallLocallyRepository",
       ":runtime:publishJvmPublicationToInstallLocallyRepository",
       ":runtime:publishJsPublicationToInstallLocallyRepository",
-      ":drivers:android-driver:publishAllPublicationsToInstallLocallyRepository",
       ":drivers:sqlite-driver:publishAllPublicationsToInstallLocallyRepository",
       ":drivers:jdbc-driver:publishAllPublicationsToInstallLocallyRepository",
       ":sqlite-migrations:publishAllPublicationsToInstallLocallyRepository",
@@ -94,13 +97,7 @@ tasks.named('test') {
         ":runtime:publishMacosX64PublicationToInstallLocallyRepository",
     )
   }
-  useJUnit {
-    if (project.hasProperty("Instrumentation")) {
-      includeCategories 'app.cash.sqldelight.Instrumentation'
-    } else {
-      excludeCategories 'app.cash.sqldelight.Instrumentation'
-    }
-  }
+  useJUnit()
   environment("ORG_GRADLE_PROJECT_sqldelightVersion", project.version)
 }
 
@@ -142,6 +139,39 @@ tasks.named("grammarkitTest") {
       ":sqldelight-compiler:environment:publishAllPublicationsToInstallLocallyRepository",
       ":sqlite-migrations:environment:publishAllPublicationsToInstallLocallyRepository",
   )
+  javaLauncher = javaToolchains.launcherFor {
+    languageVersion = JavaLanguageVersion.of(17)
+  }
+  environment("ORG_GRADLE_PROJECT_sqldelightVersion", project.version)
+}
+
+tasks.named("instrumentationTest") {
+  dependsOn(
+      ":drivers:android-driver:publishAllPublicationsToInstallLocallyRepository",
+      ":dialects:sqlite-3-18:publishAllPublicationsToInstallLocallyRepository",
+      ":sqldelight-compiler:dialect:publishAllPublicationsToInstallLocallyRepository",
+      ":runtime:publishKotlinMultiplatformPublicationToInstallLocallyRepository",
+      ":runtime:publishJvmPublicationToInstallLocallyRepository",
+      ":runtime:publishJsPublicationToInstallLocallyRepository",
+      ":drivers:sqlite-driver:publishAllPublicationsToInstallLocallyRepository",
+      ":drivers:jdbc-driver:publishAllPublicationsToInstallLocallyRepository",
+      ":sqlite-migrations:publishAllPublicationsToInstallLocallyRepository",
+      ":sqldelight-compiler:publishAllPublicationsToInstallLocallyRepository",
+      ":sqldelight-gradle-plugin:publishAllPublicationsToInstallLocallyRepository",
+      ":sqldelight-compiler:environment:publishAllPublicationsToInstallLocallyRepository",
+      ":sqlite-migrations:environment:publishAllPublicationsToInstallLocallyRepository",
+  )
+  if (HostManager.hostIsMac) {
+    dependsOn(
+      ":drivers:native-driver:publishIosX64PublicationToInstallLocallyRepository",
+      ":drivers:native-driver:publishIosSimulatorArm64PublicationToInstallLocallyRepository",
+      ":drivers:native-driver:publishKotlinMultiplatformPublicationToInstallLocallyRepository",
+      ":runtime:publishIosX64PublicationToInstallLocallyRepository",
+      ":runtime:publishIosSimulatorArm64PublicationToInstallLocallyRepository",
+      ":runtime:publishIosArm64PublicationToInstallLocallyRepository",
+      ":runtime:publishMacosX64PublicationToInstallLocallyRepository",
+    )
+  }
   javaLauncher = javaToolchains.launcherFor {
     languageVersion = JavaLanguageVersion.of(17)
   }
