+++ sqldelight-gradle-plugin/src/instrumentationTest/kotlin/app/cash/sqldelight/TestUtil.kt
@@ -1,0 +1,3 @@
+package app.cash.sqldelight
+
+internal fun String.withInvariantPathSeparators() = replace("\\", "/")
