+++ sqldelight-gradle-plugin/src/instrumentationTest/kotlin/app/cash/sqldelight/WithCommonConfiguration.kt
@@ -7,8 +7,12 @@ internal fun GradleRunner.withCommonConfiguration(projectRoot: File): GradleRunn
   File(projectRoot, "gradle.properties").writeText(
     """
       |org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=1g
+      |android.useAndroidX=true
       |
     """.trimMargin(),
   )
+  File(projectRoot, "local.properties").apply {
+    if (!exists()) writeText("sdk.dir=${androidHome()}\n")
+  }
   return withProjectDir(projectRoot)
 }
