+++ sqldelight-gradle-plugin/src/instrumentationTest/kotlin/app/cash/sqldelight/integrations/KotlinVersionsTest.kt
@@ -37,13 +37,13 @@ class KotlinVersionsTest(val kotlinVersion: String) {
   }
 
   @Test
-  fun `integration jvm compiles successfully with different Kotlin versions`() {
-    val integrationRoot = File("src/test/integration")
+  fun `kotlin mpp compiles successfully with different Kotlin versions`() {
+    val integrationRoot = File("src/test/kotlin-mpp")
 
     val runner = GradleRunner.create()
       .forwardOutput()
       .withCommonConfiguration(integrationRoot)
-      .withArguments("clean", "compileKotlin", "--stacktrace", "-PoverwriteKotlinVersion=$kotlinVersion")
+      .withArguments("clean", "assemble", "--stacktrace", "-PoverwriteKotlinVersion=$kotlinVersion")
 
     val result = runner.build()
     assertThat(result.output).contains("BUILD SUCCESSFUL")
