--- sqldelight-gradle-plugin/src/test/kotlin/app/cash/sqldelight/Instrumentation.kt
@@ -1,3 +1,0 @@
-package app.cash.sqldelight
-
-interface Instrumentation
