+++ sqldelight-gradle-plugin/src/test/kotlin/app/cash/sqldelight/integrations/IntegrationTest.kt
@@ -15,13 +15,11 @@
  */
 package app.cash.sqldelight.integrations
 
-import app.cash.sqldelight.Instrumentation
 import app.cash.sqldelight.withCommonConfiguration
 import com.google.common.truth.Truth.assertThat
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
 import org.junit.Test
-import org.junit.experimental.categories.Category
 import java.io.File
 
 class IntegrationTest {
@@ -144,97 +142,6 @@ class IntegrationTest {
   }
 
   @Test
-  @Category(Instrumentation::class)
-  fun integrationTestsAndroid() {
-    val runner = GradleRunner.create()
-      .withCommonConfiguration(File("src/test/integration-android"))
-      .withArguments("clean", "connectedCheck", "--stacktrace")
-
-    val result = runner.build()
-    assertThat(result.output).contains("BUILD SUCCESSFUL")
-  }
-
-  @Test
-  @Category(Instrumentation::class)
-  fun integrationTestsAndroidVariants() {
-    val runner = GradleRunner.create()
-      .withCommonConfiguration(File("src/test/integration-android-variants"))
-      .forwardOutput()
-
-    val generateDebugResult = runner
-      .withArguments("clean", "generateDebugQueryWrapperInterface")
-      .build()
-    assertThat(generateDebugResult.output).contains("BUILD SUCCESSFUL")
-
-    val generateReleaseResult = runner
-      .withArguments("generateReleaseQueryWrapperInterface")
-      .build()
-    assertThat(generateReleaseResult.output).contains("BUILD SUCCESSFUL")
-
-    val testDebugResult = runner
-      .withArguments("testDebugUnitTest")
-      .build()
-    assertThat(testDebugResult.output).contains("BUILD SUCCESSFUL")
-    assertThat(
-      requireNotNull(
-        testDebugResult.task(":generateDebugQueryWrapperInterface"),
-        {
-          "Could not find task in ${testDebugResult.tasks}"
-        },
-      ).outcome,
-    ).isEqualTo(TaskOutcome.UP_TO_DATE)
-
-    val testReleaseResult = runner
-      .withArguments("testReleaseUnitTest")
-      .build()
-    assertThat(testReleaseResult.output).contains("BUILD SUCCESSFUL")
-    assertThat(
-      requireNotNull(
-        testReleaseResult.task(":generateReleaseQueryWrapperInterface"),
-        {
-          "Could not find task in ${testDebugResult.tasks}"
-        },
-      ).outcome,
-    ).isEqualTo(TaskOutcome.UP_TO_DATE)
-  }
-
-  @Test
-  @Category(Instrumentation::class)
-  fun integrationTestsAndroidLibrary() {
-    val integrationRoot = File("src/test/integration-android-library")
-
-    // Copy the normal android integration files over.
-    val target = File(integrationRoot, "src")
-    File(File("src/test/integration-android"), "src").copyRecursively(target)
-    try {
-      val runner = GradleRunner.create()
-        .withCommonConfiguration(integrationRoot)
-        .withArguments("clean", "connectedCheck", "--stacktrace")
-
-      val result = runner.build()
-      assertThat(result.output).contains("BUILD SUCCESSFUL")
-    } finally {
-      target.deleteRecursively()
-    }
-  }
-
-  @Test
-  @Category(Instrumentation::class)
-  fun `integration test android target of a multiplatform project`() {
-    val integrationRoot = File("src/test/integration-multiplatform")
-    val buildGradle = File(integrationRoot, "build.gradle").apply { deleteOnExit() }
-    File(integrationRoot, "android-build.gradle").copyTo(buildGradle, overwrite = true)
-
-    val runner = GradleRunner.create()
-      .withCommonConfiguration(integrationRoot)
-      .withArguments("clean", "test", "--stacktrace")
-
-    val result = runner.build()
-    assertThat(result.output).contains("BUILD SUCCESSFUL")
-  }
-
-  // @Category(Instrumentation::class)
-  @Test
   fun `integration test multiplatform project with async generated code`() {
     val integrationRoot = File("src/test/integration-multiplatform-async")
 
@@ -242,38 +149,6 @@ class IntegrationTest {
       .withCommonConfiguration(integrationRoot)
       .withArguments("clean", "test", "--stacktrace")
       .withDebug(true)
-
-    val result = runner.build()
-    assertThat(result.output).contains("BUILD SUCCESSFUL")
-  }
-
-  @Test
-  @Category(Instrumentation::class)
-  fun `integration test ios target of a multiplatform project`() {
-    val integrationRoot = File("src/test/integration-multiplatform")
-    val buildGradle = File(integrationRoot, "build.gradle").apply { deleteOnExit() }
-    File(integrationRoot, "ios-build.gradle").copyTo(buildGradle, overwrite = true)
-
-    val runner = GradleRunner.create()
-      .forwardOutput()
-      .withCommonConfiguration(integrationRoot)
-      .withArguments("clean", "iosTest", "--stacktrace")
-
-    val result = runner.build()
-    assertThat(result.output).contains("BUILD SUCCESSFUL")
-  }
-
-  @Test
-  @Category(Instrumentation::class)
-  fun `integration metadata task compiles successfully`() {
-    val integrationRoot = File("src/test/integration-multiplatform")
-    val buildGradle = File(integrationRoot, "build.gradle").apply { deleteOnExit() }
-    File(integrationRoot, "ios-build.gradle").copyTo(buildGradle, overwrite = true)
-
-    val runner = GradleRunner.create()
-      .forwardOutput()
-      .withCommonConfiguration(integrationRoot)
-      .withArguments("clean", "compileKotlinMetadata", "--stacktrace")
 
     val result = runner.build()
     assertThat(result.output).contains("BUILD SUCCESSFUL")
